Rename mapping.json from requires. - Renamed mapping.json to mapping_100.json from all instances of requires of indexes/repo/mapping.json
